+++ team_AvangardTeenOrder/src/main/java/lv/avangardteen/core/request/ChangePersonalSizeRequest.java
@@ -1,13 +1,13 @@
 package lv.avangardteen.core.request;
 
 public class ChangePersonalSizeRequest {
-    public long id;
+    public int id;
     public int pelvisWidth; //ширина таза
     public int thighLength; //длинна бедра
     public int backHeight; //высота спины
     public int shinLength;
 
-    public ChangePersonalSizeRequest(long id, int pelvisWidth, int thighLength, int backHeight, int shinLength) {
+    public ChangePersonalSizeRequest(int id, int pelvisWidth, int thighLength, int backHeight, int shinLength) {
         this.id = id;
         this.pelvisWidth = pelvisWidth;
         this.thighLength = thighLength;
@@ -16,7 +16,7 @@ public class ChangePersonalSizeRequest {
     }
 
 
-    public long getId() {
+    public int getId() {
         return id;
     }
